Refactor: HttpRequest와 HttpResponse 생성
